Crate epic_store

Storage of core types using RocksDB.

Re-exports

pub use crate::lmdb::*;

Modules

leaf_set

The Epic leaf_set implementation. Compact (roaring) bitmap representing the set of leaf positions that exist and are not currently pruned in the MMR.

lmdb

Storage of core types using LMDB.

pmmr

Implementation of the persistent Backend for the prunable MMR tree.

prune_list

The Epic “Prune List” implementation.

types

Common storage-related types

Functions

read_bitmap

Read Bitmap from a file

save_via_temp_file

Creates temporary file with name created by adding temp_suffix to path. Applies writer function to it and renames temporary file into original specified by path.

to_key

Build a db key from a prefix and a byte vector identifier.

to_key_u64

Build a db key from a prefix and a byte vector identifier and numeric identifier

u64_to_key

Build a db key from a prefix and a numeric identifier.