Struct epic_store::leaf_set::LeafSet

pub struct LeafSet { /* fields omitted */ }

Compact (roaring) bitmap representing the set of positions of leaves that are currently unpruned in the MMR.

Implementations

impl LeafSet

pub fn open<P: AsRef<Path>>(path: P) -> Result<LeafSet>

Open the remove log file. The content of the file will be read in memory for fast checking.

pub fn copy_snapshot<P: AsRef<Path>>(path: P, cp_path: P) -> Result<()>

Copies a snapshot of the utxo file into the primary utxo file.

pub fn removed_pre_cutoff(
&self,
cutoff_pos: u64,
rewind_rm_pos: &Bitmap,
prune_list: &PruneList
) -> Bitmap

Calculate the set of pruned positions up to and including the cutoff_pos. Uses both the leaf_set and the prune_list to determine prunedness.

pub fn rewind(&mut self, cutoff_pos: u64, rewind_rm_pos: &Bitmap)

Rewinds the leaf_set back to a previous state. Removes all pos after the cutoff. Adds back all pos in rewind_rm_pos.

pub fn add(&mut self, pos: u64)

Append a new position to the leaf_set.

pub fn remove(&mut self, pos: u64)

Remove the provided position from the leaf_set.

pub fn snapshot(&self, header: &BlockHeader) -> Result<()>

Saves the utxo file tagged with block hash as filename suffix. Needed during fast-sync as the receiving node cannot rewind after receiving the txhashset zip file.

pub fn flush(&mut self) -> Result<()>

Flush the leaf_set to file.

pub fn discard(&mut self)

Discard any pending changes.

pub fn includes(&self, pos: u64) -> bool

Whether the leaf_set includes the provided position.

pub fn len(&self) -> usize

Number of positions stored in the leaf_set.

pub fn is_empty(&self) -> bool

Is the leaf_set empty.

pub fn iter(&self) -> impl Iterator<Item = u64> + '_

Iterator over positionns in the leaf_set (all leaf positions).