Struct epic_store::pmmr::PMMRBackend

pub struct PMMRBackend<T: PMMRable> { /* fields omitted */ }

PMMR persistent backend implementation. Relies on multiple facilities to handle writing, reading and pruning.

A main storage file appends Hash instances as they come. This AppendOnlyFile is also backed by a mmap for reads.

An in-memory backend buffers the latest batch of writes to ensure the PMMR can always read recent values even if they haven’t been flushed to disk yet.

A leaf_set tracks unpruned (unremoved) leaf positions in the MMR..

A prune_list tracks the positions of pruned (and compacted) roots in the MMR.

Implementations

impl<T: PMMRable> PMMRBackend<T>

pub fn new<P: AsRef<Path>>(
data_dir: P,
prunable: bool,
version: ProtocolVersion,
header: Option<&BlockHeader>
) -> Result<PMMRBackend<T>>

Instantiates a new PMMR backend. If optional size is provided then treat as “fixed” size otherwise “variable” size backend. Use the provided dir to store its files.

pub fn unpruned_size(&self) -> u64

Number of hashes in the PMMR stored by this backend. Only produces the fully sync’d size.

pub fn data_size(&self) -> u64

Number of elements in the underlying stored data. Extremely dependent on pruning and compaction.

pub fn hash_size(&self) -> u64

Size of the underlying hashed data. Extremely dependent on pruning and compaction.

pub fn sync(&mut self) -> Result<()>

Syncs all files to disk. A call to sync is required to ensure all the data has been successfully written to disk.

pub fn discard(&mut self)

Discard the current, non synced state of the backend.

pub fn check_compact(
&mut self,
cutoff_pos: u64,
rewind_rm_pos: &Bitmap
) -> Result<bool>

Takes the leaf_set at a given cutoff_pos and generates an updated prune_list. Saves the updated prune_list to disk, compacts the hash and data files based on the prune_list and saves both to disk.

A cutoff position limits compaction on recent data. This will be the last position of a particular block to keep things aligned. The block_marker in the db/index for the particular block will have a suitable output_pos. This is used to enforce a horizon after which the local node should have all the data to allow rewinding.